many chagnes to make things actually function

+++ src/main/java/frc/robot/Constants.java
@@ -28,8 +28,8 @@ public final class Constants {
     public static final int           HANGER_LEFT_MOTOR_ID        = 10;
     public static final int           HANGER_RIGHT_MOTOR_ID       = 11;
  
-    public static final int           LOOP_TIME_MS                = 20;
-    public static final int           LOOP_TIME_SECONDS           = LOOP_TIME_MS / 1000;
+    public static final double        LOOP_TIME_MS                = 20;
+    public static final double        LOOP_TIME_SECONDS           = LOOP_TIME_MS / 1000;
 
 
     public static final String        LIMELIGHT_NAME              = "limelight";
